// types/CarDetails.tsx
export interface CarDetails {
  Make: string;
  Model: string;
  "Model Year": string;
  Trim: string;
  BodyClass: string;
  DriveType: string;
  EngineCylinders: string;
  "Fuel Type Primary": string;
  GVWR: string;
  TransmissionStyle: string;
  VIN: string;
  ManufacturerName: string;
  PlantCity: string;
  PlantCountry: string;
  PlantState: string;
  VehicleType: string;
  DisplacementL: string;
  EngineConfiguration: string;
  FuelDeliveryType: string;
  SeatBeltsType: string;
  AirBagLocations: string;
  ErrorCode: string;
  ErrorText: string;
  SuggestedVIN: string;
  PossibleValues: string;
  AdditionalErrorText: string;
  VehicleDescriptor: string;
  DestinationMarket: string;
  Series: string;
  PlantCompanyName: string;
  Trim2: string;
  Series2: string;
  Note: string;
  BasePrice: string;
  NonLandUse: string;
  Doors: string;
  Windows: string;
  WheelBaseType: string;
  TrackWidth: string;
  GrossVehicleWeightRatingFrom: string;
  BedLength: string;
  CurbWeight: string;
  WheelBaseFrom: string;
  WheelBaseTo: string;
  GrossCombinationWeightRatingFrom: string;
  GrossCombinationWeightRatingTo: string;
  GrossVehicleWeightRatingTo: string;
  BedType: string;
  CabType: string;
  TrailerTypeConnection: string;
  TrailerBodyType: string;
  TrailerLength: string;
  OtherTrailerInfo: string;
  NumberOfWheels: string;
  WheelSizeFront: string;
  WheelSizeRear: string;
  EntertainmentSystem: string;
  SteeringLocation: string;
  NumberOfSeats: string;
  NumberOfSeatRows: string;
  TransmissionSpeeds: string;
  Axles: string;
  AxleConfiguration: string;
  BrakeSystemType: string;
  BrakeSystemDescription: string;
  OtherBatteryInfo: string;
  BatteryType: string;
  NumberOfBatteryCellsPerModule: string;
  BatteryCurrentFrom: string;
  BatteryVoltageFrom: string;
  BatteryEnergyFrom: string;
  EVDriveUnit: string;
  BatteryCurrentTo: string;
  BatteryVoltageTo: string;
  BatteryEnergyTo: string;
  NumberOfBatteryModulesPerPack: string;
  NumberOfBatteryPacksPerVehicle: string;
  ChargerLevel: string;
  ChargerPower: string;
  DisplacementCC: string;
  DisplacementCI: string;
  EngineStrokeCycles: string;
  EngineModel: string;
  EnginePower: string;
  ValveTrainDesign: string;
  FuelTypeSecondary: string;
  EngineBrakeFrom: string;
  CoolingType: string;
  EngineBrakeTo: string;
  ElectrificationLevel: string;
  OtherEngineInfo: string;
  Turbo: string;
  TopSpeed: string;
  EngineManufacturer: string;
  Pretensioner: string;
  OtherRestraintSystemInfo: string;
  CurtainAirBagLocations: string;
  SeatCushionAirBagLocations: string;
  FrontAirBagLocations: string;
  KneeAirBagLocations: string;
  SideAirBagLocations: string;
  AntiLockBrakingSystem: string;
  ElectronicStabilityControl: string;
  TractionControl: string;
  TirePressureMonitoringSystemType: string;
  ActiveSafetySystemNote: string;
  AutoReverseSystem: string;
  AutomaticPedestrianAlertingSound: string;
  EventDataRecorder: string;
  KeylessIgnition: string;
  SAEAutomationLevelFrom: string;
  SAEAutomationLevelTo: string;
  AdaptiveCruiseControl: string;
  CrashImminentBraking: string;
  BlindSpotWarning: string;
  ForwardCollisionWarning: string;
  LaneDepartureWarning: string;
  LaneKeepingAssistance: string;
  BackupCamera: string;
  ParkingAssist: string;
  BusLength: string;
  BusFloorConfigurationType: string;
  BusType: string;
  OtherBusInfo: string;
  CustomMotorcycleType: string;
  MotorcycleSuspensionType: string;
  MotorcycleChassisType: string;
  OtherMotorcycleInfo: string;
  DynamicBrakeSupport: string;
  PedestrianAutomaticEmergencyBraking: string;
  AutomaticCrashNotification: string;
  DaytimeRunningLight: string;
  HeadlampLightSource: string;
  SemiautomaticHeadlampBeamSwitching: string;
  AdaptiveDrivingBeam: string;
  RearCrossTrafficAlert: string;
  RearAutomaticEmergencyBraking: string;
  BlindSpotIntervention: string;
  LaneCenteringAssistance: string;
}